ChatWIthPDF/requirements.txt
langchain
langchain-community
pypdf
langchain-chroma
langchain-google-genai
streamlit
python-dotenv